requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
cssselect==1.2.0
d==0.2.2
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==25.8.0
isort==5.13.2
lxml==5.2.2
Markdown==3.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
Pygments==2.18.0
PyJWT==1.7.1
pyquery==2.0.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
